Move debug related gems out of gemspec.
When doing development, running `bundle install` will take into account
the Gemfile. By default the Gemfile references the gemspec so everything
appears to come from the gemspec alone. We are free to add more
dependencies to the Gemfile as we see fit.

Gems added directly to the Gemfile will **not** be included in the
dependencies list when the gem is built.