Update RSpec configuration to latest
Most of the settings are the same but this includes the out of the doc
documentation which accompanies the settings. This makes it easier for
use to understand what a setting does when we review the file later.

Since some of the settings are specific to the newer RSpec this bumps
the enforced version in the gemspec.