Make specs happier in isolation
A lot of specs have a hidden implicit dependency that the configuration
is set prior to them running. While this probably does more harm by
hiding the fact things aren't set, it makes more sense to lazy load the
configuration via the reader than the block setter. It also gives us a
solid place to put defaults in the future.