Re-write configuration
This intent here is to create a more structured configuration. I wanted
a clear way to self document each setting and any default values they
may have. I also wanted the ability to easily extend this in the future
knowing that we do have several additional settings we want to add.

I am purposefully avoiding using `Struct` for the following reasons:

  - multiple accessors through `[]` and setting names make it difficult
    to customize access when necessary
  - equality based on value

    `Configuration` is a mutable container of settings. I do not view
    them as value objects. Therefore, their equality shouldn't be based
    on their data values either.
  - customization is still required for built-in default values
  - cannot use `super` for reader / writer delegation when customizing

This new design is more complex. But it provides several improvements:

  - hides explicit access of ivars through messages
  - allows customizing readers with normal call chain behavior
  - self documenting what is considered a setting and their defaults
  - enforces set of settings through messages rather than maintaining an
    additional mutable list (requiring both to be kept in sync)

    This is arguably much slower. A quick benchmark says it's around
    four times slower. The intent is not to be reading the list of
    settings often. We do use it in the hash (`to_h`) conversion but we
    don't expect that to be called often. If it is, we could probably
    memoize the set in some way while still retaining the benefits of
    using `Settings` to hold the official list.

Regarding having `Settings` only include readers. Originally I named
this `Readers` following the convention RSpec set forth. When looking at
the ancestor chain it reads nicely as "oh that's where readers are",
however, that really doesn't express the intent. A customized reader may
be on `Configuration` so `Readers` is misleading. It really holds the
names of the defined `Settings`.

I went back and forth a bit on placing the writers directly on
`Configuration`. This prevents us from delegating up the call chain and
additionally requires us to explicitly set the ivar when we customize
it. But including the writers in `Settings` means we have to parse
through what's a reader and a writer defeating the purpose of the
module. I thought about then including a `Writers` module, but at this
point we don't need it and that seemed too much (as if it already isn't
too much for a simple configuration :grimacing:). If I did that would I
then want a `Defaults` module too? I saw things quickly getting out of
control so I decided to just keep it simple to what we need now.

The macro which defines the settings could alternatively be written in
less lines of code as the following:

```ruby
def self.add_setting(name, default: nil)
  attr_writer name
  Settings.module_exec { attr_reader name }
  define_method(name) { super() || default } if default
end
```

The reason it is not is the explicit use of Weirich-style blocks.

> NOTE: For those who have not heard of Weirich-style blocks.
>
> This style of blocks, and method calls, is named after the late Jim
> Weirich. It is so named because the main block style conventions were
> originally based around single vs multi-line; brackets for single line
> `do`/`end` for multi-line. While this seems logical on the face of it,
> it tends to feel _weird_ with method changing:
>
> ```ruby
> array.each { |element| do_something element }
>
> array.each do |element|
>   element = normalize(element)
>   do_something element
> end
>
> array.select do |element|
>        # several steps
>        valid? element
>      end
>      .map { |element| transform element }
>
> # This was more of an issue in Ruby 1.8 which required a trailing dot
> array.select do |element|
>        # several steps
>        valid? element
>      end.
>      map { |element| transform element }
> ```
>
> Jim proposed an alternative provided more information to a developer
> by attaching semantic mean each style. This style stated to use
> brackets for blocks which provided a meaningful return value and use
> `do`/`end` for blocks which simply produce side-effects. This same
> style applied to method calls and parentheses.
>
> Using the same examples above would look like:
>
> ```ruby
> array.each do |element|
>   do_something element
> end
>
> array.each do |element|
>   element = normalize(element)
>   do_something element
> end
>
> array.select { |element|
>        # several steps
>        valid?(element)
>      }
>      .map { |element| transform(element) }
>
> # With Ruby 1.8 trailing dot
> array.select { |element|
>        # several steps
>        valid?(element)
>      }.
>      map { |element| transform(element) }
> ```

New Setting: `doc_root`
-----------------------

We allow configuration nearly everything related to the repo. Aside from
the Github URL, which we control due to behavioral differences between
the various domains, the only other thing which isn't configurable is
where in the repo we look for the docs. We had a convention of "docs"
for our repos but there's no reason to force that. This new setting
allows that to be changed, while still defaulting to this convention.

Breaking Change: Remove setting inference from `ENV`
----------------------------------------------------

Magically inferring settings from the environment seems like a good
idea. There is much precedence for this style from several common gems.
However, with limited exception (I'm looking at you database and general
environment name settings) this is a bad idea. Some of the main pitfalls
with this design are:

- Name conflicts across gems
- Requires potential duplicating of settings for similar behavior
- Difficult debugging when combined with explicit settings (e.g. I'm not
  setting `xyz` in the config block. Why is it set to `wtf`?)
- Doesn't scale to all types of settings nor work particularly well with
  a lot of settings.

To avoid the first issue we used the verbose prefixed of `DOC_REPO_` to
differentiate the settings. However, this means if someone includes a
Github API gem in their app/program which should point to the same place
(maybe the doc site also includes releases or other similar content) the
settings may now be duplicated in two environment variables. As with any
duplication things must be kept in sync between the two.

With magic environment variables there's nothing you can do to prevent
the third issue. You just have to document them well and hope people are
aware. This is why they work best for a small number (fourth issue) of
extremely well known settings.

Essentially this all boils down by us choosing to magically infer
settings from the environment we've imposed this design decision on all
developers. This increases their maintenance and cognitive costs. I'm
not saying such a design is bad, I'm saying we should leave it up to the
developer to decide to use it. While it was a standard in the past it is
slowing becoming less of a forced design by gems. Rails recently
announced encrypted secrets to avoid putting things in the environment
except for a master key.