Rewrite repository
The existing UX for working with a repository felt a bit odd. Part of
this is because the names of things imply a more format based response
than the actual behavior which occurs (i.e. `redirect` is not a format).
Another main issue is that "errors" bubble up to the surface requiring
rescuing. This control-flow-by-exception strategy is non-idea as the
common "Not Found" behavior is not exceptional. Compound this is the
fact that all remote HTTP errors are considered as `Not Found` results.
Even though the underlying error is exposed via a reader this limits the
ability to cleanly provide different error handling strategies.

To address these concerns the goals of this re-write are as follows:

- Rename things to improve implied intent
- Treat HTTP errors as a normal part of the flow
- Improve defined responsibilities of the objects involved

First up the general interface has been renamed from `respond_with` to
`request`. With `respond_with` seems good on the surface it is
overloaded by Rails' `respond_to`/`respond_with` which implies content
type variability. However, our behavior is closer to handling different
request statuses (e.g. redirect). As we talk to a remote server over
HTTP it makes sense to call our interface a `request`.

I also considered the following names:

  - fetch: verb. go for and then bring back

    I chose not to use this because it implies returning something and
    we may not return anything such as in the case of a redirect.

  - retrieve: verb. get or bring (something) back

    I decided against this for the same reasons as `fetch`

  - find: verb. recognize or discover (something) to be present

    I decided against this for the same reasons as `fetch`.
    Additionally, this implies an interface similar to
    `ActiveRecord#find` and I wanted to avoid that overloading.

  - detect: verb. discover or identify the presence or existence of

    I was originally leaning toward this, but after a while it just
    didn't sound right given this higher level context.

  - search: verb. try to find something by looking or otherwise seeking
    carefully and thoroughly

    I strongly considered this but it felt wrong as generally we
    "search" for a collection of things and we are only working with a
    single page here.

Next I wanted to define the responsibilities of the `Repository`:

  - handling requests for a page with respect to a Github repository
  - implementing default fallback handling behavior
  - generating a URI for a page slug
  - understanding the general backing API structure (required for URI
    generation); this has an implicit responsibility of knowing the host
    domain as that generally defines the available API structure

Without an `org` or `repo` the term "repository" doesn't mean much. In
Github world a "repository" is defined by those two qualifiers. For our
purposes we include `branch`, which determines the "view" of a
repository, as a defining piece of information. Thus a `Repository`
requires a `Configuration` in order to be created.

While it is the responsibility of the repository to understand the API,
it is not it's responsibility to handle the actual lower level call. Nor
processing of the response. Additionally, looking to the future an HTTP
level cache is not part of these responsibilities either.

To ease implementation and facilitate easier changes later this allows
the lower level HTTP adapter to be injected. Since we don't want to
require callers to know how to generate this we provide a convenient
default.

As the repository is responsible for understanding the backing API
structure it has an implicit responsibility for knowing which assets to
render and which to redirect for raw access. This inverts the logic
redirecting everything which is not a known formattable type. This
provides greater flexibility for what can be stored within documentation
(i.e. videos, zipped tools, etc). Since it's likely people may want to
serve other documentation formats we expose this setting through
`DocRepo::Configuration#doc_formats`. Similarly, the fallback extension
for when one isn't provided in the page slug (previously called the
"default" but really we "fall back" to using it when processing) is now
also configurable through `DocRepo::Configuration#fallback_ext`.

Since the repository is responsible for knowing what to redirect it
handles that during the resource detection process. For things it deems
as possible documentation pages the actual retrieval is delegated to the
underlying HTTP adapter.

Whatever the response from the retrieval it calls the appropriate
handler. The repository is responsible for sensible handler fall back
strategies; we can't let something not be handled b/c it likely
indicates a developer error and could cause difficult to debug
downstream issues. For redirects and successful retrievals we throw an
error when there is no handler. For example, in Rails if nothing is
performed here it's likely a view or missing view error will result. As
we can't clearly define a sensible fall back we raise
`DocRepo::UnhandledAction` to better inform the developer of their
oversight.

However, for errors we do have some sensible alternatives. For documents
which are missing (i.e. not found) it makes sense to fall back to the
general error handling strategy. In the absence of an error strategy a
good fall back is to raise the error encountered.

To flush out our behavioral logic we include some HTTP adapter and
result class stubs. We fully intend to replace this with more
appropriate implementations later.

New Features
------------

- Provide a more helpful error through `DocRepo::UnhandledAction` when a
  required handler (e.g. `redirect` or `complete`) is missing

- Treat result handlers `not_found` and `error` as optional

  The fallback strategy for `not_found` is to call the `error` handler.
  When the `error` is not defined the error is raised.

- New configuration setting `doc_formats`

  This allows custom changes to the list of which formats are served and
  which are redirected to the raw resource/asset.

- New configuration setting `fallback_ext`

  For slugs without an extension (think index type pages) this allows
  customization of what's applied by default. This was added as some
  people prefer to use `.markdown` as the extension over `.md`. But
  people could also very well expect a `.html` as their standard.

Breaking Changes
----------------

- `DocRepo#respond_with` is now `DocRepo#request` and the object yielded
  to the block uses the following syntax:

  ```ruby
  DocRepo.request(params[:slug]) do |on|
    on.redirect do |url|
    end

    on.complete do |doc|
    end

    on.not_found do |error|
    end

    on.error do |error|
    end
  end
  ```

- `DocRepo::NotFound` has been replaced by a more generic
  `DocRepo::HttpError` which inherits from `DocRepo::Error`

  Since `Not Found` responses are a common error there is the predicate
  query `DocRepo::HttpError#not_found?` available for use.

  `DocRepo::Error` does **not** provide a `base` reader exposing wrapped
  errors. Instead it's subclasses may encapsulated a lower level error.

- `DocRepo::Repository` requires a `DocRepo::Configuration` on creation