Add new cache configuration settings
The goal here is to be able to provide a simplistic HTTP cache layer.
This is intended for both speed and reducing rate limit issues (though
with the new domain that is almost a non-issue).

At this time we aren't sure exactly how we are going to use this but we
do know that we are going to follow the common Rails cache store
conventions of `fetch` and `write`. We are not specifying `read` here
simply because our intended use will be to perform the remote request
when there is no local cache.

As we cannot guarantee what cache stores are available this defaults to
a null cache. As this performs no actual storage there is only ever a
need for a single instance. To enforces this, which is something I
generally avoid but here it really makes sense, we make this a
singleton. We do assume that cache stores take a options as a hash.