Add configurable handler for doc requests
The idea here is to improve the document response block handling. In
light of the recent rate limit `403 Forbidden` issue we want to be able
to distinguish between real errors like those and normal `404 Not Found`
pages. Additionally, we want to transition away from our exception
strategy because encountering errors isn't really exceptional with
respect to HTTP requests.

The basic idea is to have an implementation like the following:

```ruby
DocRepo.respond(params[:slug]) do |on|
  on.complete do
  end

  on.redirect do
  end

  on.not_found do
  end

  on.error do
  end
end
```

Issues with Current Design
--------------------------

- Extending functionality means creating both new factories and type
  switched method implementations (a macro could ease this)

- There's no ability to have fallback behavior due to the typed nature
  of the implementation and the fact methods are called directly

  For example, we'd ultimately like the not found handler to be
  optional. When it isn't defined the error handler should be called
  instead. We'd like to even make that optional falling back to raising
  the errors as a final behavior. But the current design makes it
  impossible to know "what's defined":

  ```ruby
  DocRepo.respond(params[:slug]) do |f|
    f.html { }
  end
  ```

  Since the actual methods are called from the block "undefined" simply
  means "not called". Thus a not found error either can't be handled or
  the handler needs to keep internal state about what other methods were
  already called to avoid the case of both the not found and general
  error methods being called.

- The existing naming of `html` conflates Rails' respond to format
  syntax with general documentation page availability handling

New Design
----------

This takes a note from the Rails' respond to format object structure.
The container, which will be yielded to the block, has no handlers
defined. Instead when a handler is defined it is stored for access
later. By storing the blocks it makes it possible to have various
fallback strategies later. It also trades the "type checks" for a
"lookup".

The internal map of the handlers is kept private to prevent undefined
handlers being added. The Rails response container is a bit more robust
in it allows _some_ dynamic formats to be added. At this time such a
feature is out of scope but is something to look to in the future.
Instead the plan is to allow `[]` and `fetch` as the interfaces for
requesting a handler. This will allow for fall back strategies as
follows:

```ruby
handler.fetch(:not_found) {
  handler.fetch(:error) { }
}
```

For improved UX a macro is defined to ensure the handler is provided a
block. Without it there's no definition of behavior and is considered a
developer error.