Implement basic naive HTTP cache
This is about as basic a cache layer as we can get. We take the provided
cache store and store the raw HTTP response into the cache. The provided
cache options are passed directly to the cache without modification.

In an attempt to avoid name collisions the cache key is composed from
the hostname and the URI. The hostname serves as an internal, to
`NetHttpAdapter`, namespace to avoid conflicts with the same URI across
different systems. In this context the term "namespace" is used purely
to convey the general concept and does not imply the hostname is
provided as a `namespace` option to the cache. As stated above the
`cache_options` are passed directly to the cache without modification.

Assumptions and Naiveté
-----------------------

We have an implicit assumption here that we can store any type of object
in the cache. This is a very naive assumption, but it keeps the
implementation here simple. This assumption will hold for general
in-memory object stores, as well as with Rails 5 cache stores (which
handle the object marshalling).

With this naive implementation the resource stored at the URI is only
fetched from the remote host when it is not in the cache. Whatever the
response may be it will then be cached and returned by future requests.
The length of time this response remains in the cache is based purely
on the `cache_options` and default settings of the cache itself.

As per the TODO comment there are some issues with this. It's possible
the remote host does not want the response cached. Or that the response
should only be valid for a limited time because the host has it's own
caching opinions (i.e. cache control headers). This implementation
ignores all of that and simply treats the cache as valid for as long as
the cache store does.

Another naive behavioral choice is caching 4xx ("client" errors) and 5xx
("server" errors) responses for the same length of time as successful
retrievals. This is mostly to keep things simple at the moment. We
assume that the host app has the ability to purge / selectively clean
cache on demand. We also assume that even a 5xx may count against a rate
limit quota. At this time we've chosen to prioritize implementation
simplicity and reducing remote calls to minimize rate limit impacts over
a more robust and complicated response cache handling process.