Implement conditional GET / HTTP cache expiration
This improves the HTTP cache layer logic by making it aware of
conditional `GET` and the basic cache expiration header. The goal here
is to balance respecting the server's stated cache behavior with
limiting overall network calls / rate limiting. Additionally, this is
only an incremental steps towards improving the HTTP cache behavior and
targets the most "simplistic" headers.

This views the following headers as most "simplistic":

- `Expires`
- `ETag` / `If-None-Match`
- `Last-Modified` / `If-Modified-Since`
- `Date`

Expiration
----------

Per the RFC 7234 regarding HTTP/1.1 Caching the main "freshness" headers
are `Expires` and `Cache-Control`. The `Expires` header _should_ be a
simple HTTP date value. Whereas the `Cache-Control` header may be
composed of several components, all of which need to be parsed and taken
into consideration. We are considering the `Cache-Control` header to be
to complex for this initial step.

This means two things:

- this is _NOT_ a proper HTTP cache implementation as `Expires` _must_
  be ignored when `Cache-Control` is present

- when there is no `Expires` header but there is a `Cache-Control`
  header we will _NOT_ expire the cache outside of the `cache_options`
  setting

At the moment the Github raw server we are using provides both of these
headers. From what I have seen the `Cache-Control` only contains either
`no-cache` (for not found / error responses) or `max-age` for available
content. In the case of available content with `max-age` the `Expires`
header is set to the pre-calculated value of `Date` + `max-age`. For the
`no-cache` (not found / error responses) there is no `Expires` header
set.

**This means we will cache errors potentially longer than available
content.**

We do plan on coming back to this later to properly implement
`Cache-Control` and implement a more robust response code based
freshness model. But for now we're sticking with implementation
simplicity and reduced network calls as the priority.

Refresh / Conditional `GET`
---------------------------

When a response is considered expired, based on the `Expires` header, we
make an attempt to refresh it. To try to help reduce network load and
respect rate limits we send a conditional `GET` for the refresh request.
We provide both the ETag based `If-None-Match` header and the weaker
`If-Modified-Since` header. `Net::HTTP` won't strip out `nil` headers
values so we remove them prior to sending. This also avoids the
following warning written to standard error by `Net::HTTP`:

    net/http: warning: nil HTTP header: XYZ

If the response comes back as `304 Not Modified` we need to "refresh"
the existing stored response. Working with `Net::HTTPResponse` objects
is painful. Due to it's internal implementation with sockets it can be
troublesome to manually create a response object. Additionally, it can
be problematic to set the body on an existing response. I found the that
the safest way to refresh our cache is to modify the existing response.
As a `304 Not Modified` generally does not provide a body, and we do not
care for it's status code, what we are most interested in is the
headers. So a fresh is simply an overwriting of the stored headers with
those we received "fresh".

In the case where something has changed the response will be the normal
`200 OK` with the content. For this case we simply replace the prior
stored response with this new one.

What this does **NOT** do
-------------------------

- implement proper `Cache-Control` header behavior
- allow configuration override of any of this behavior
- implement response code specific cache logic
- handle the edge case where content returns already expired (in this
  case two requests will always be issued)

See Also
--------

- [Hypertext Transfer Protocol (HTTP/1.1): Conditional Requests](https://tools.ietf.org/html/rfc7232)
- [Hypertext Transfer Protocol (HTTP/1.1): Caching](https://tools.ietf.org/html/rfc7234)
- https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control
- https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/ETag
- https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Expires
- https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-Modified-Since
- https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match
- https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Last-Modified