Drop Ruby 2.3 support
Ruby prior to 2.4 does not support `nil.dup`. While this isn't a big
deal it does make certain things more painful than necessary. There are
also some nice new Ruby 2.4 improvements such as `Hash#compact` and
`Hash#compact!` which replaces:

  - `Hash#keep_if { |_, v| v }`
  - `Hash#delete_if { |_, v| v.nil? }`

As this is a re-write we want to look towards the future and minimize
concessions for legacy implementations.