Extract HTTP result classes into separate files
At this point things feel relatively stable. At least enough for us to
move all of the HTTP result classes into separate files. This way they
can each be loaded as necessary without requiring everything be grouped
together.

As the `Doc` result class is one of the more complicated implementations
we take a moment to wrap the existing behavior in specs.