Add additional cache attributes for documents
The goal here is to provide some basic app cache helpers. These are
meant as reasonable default implementations for an app which wishes to
produce / store a cache based on the content.

It should be noted that an app cache is meant to be treated differently
than our HTTP cache in `DocRepo::NetHttpAdapter`. In general an app will
be concerned with the content (raw or formatted) and probably not with
the vast majority of headers we receive from the origin server.

For the cache key we simply use the URI. This is meant to identify the
content on the remote system so it seems like a natural choice for a
cache key. It's also important to note that in general the URI will
include an extension so there is an implicit "format" built in already.

In general the origin server will provide us an `ETag` for remote
content, but it should be noted it is _not required_ to do so. When
considering this we need to determine a fallback behavior for
`cache_version` when there is no `ETag` header. The more I thought about
this the more I realized that we need to provide our own cache version
completely separate from the remote `ETag` header.

What pushed me to this conclusion was considering the case of if today
we received an `ETag` but tomorrow the origin server stopped providing
one (or the reverse behavior). In these cases the `cache_version` would
suddenly change even though the actual `content` remained the same.

The logical next question is then what will be considered for generating
the cache version and how will we do it. I think simply put we are most
concern with the actual `content`. Things like `content_type` are more a
formatting issue, something we haven't fully addressed yet, and if the
base content doesn't change things can be formatted in a number of
different ways. Also, if a major format change did occur (such as from
one image type to another) the actual content would likely change as
well - not to mention a "format" extension is already built into the
`cache_key`.

For content related purposes we don't really care when the content was
changed. Just what it contains. For this reason (also due to the fact
that our main origin server GRC doesn't provide this) we leave this out
of the calculation as well. This really just leave the actual raw
content.

Next we need to consider a hashing/digest algorithm. Looking at Rails 5
they use `Digest::MD5`. This is a well know 16-byte algorithm. However,
it is considered "broken" for cryptographic purposes and has a
relatively high rate of collisions. Looking at the `ETag` provided by
GRC we can see it is a 20-byte hex digest. There are several hashing
algorithms which produce similar length results. One of the more well
known ones are SHA1 and RIPEMD-160.

We can't forget that we aren't trying to "hide" content. We provide the
content when we provide the cache version. So being "cryptographically"
secure isn't a priority. What we are concerned with is that there's a
low risk of collision and that we have little overhead in terms of
computation time of the digest. Taking that into account it seems like
MD5 is a good choice (after all Rails has decided to use it). Out of
curiosity we profile Ruby's implementation of these algorithms to see
how they perform (see benchmarks/digests.rb):

```
Warming up --------------------------------------
                 MD5    32.125k i/100ms
              RMD160    21.157k i/100ms
                SHA1    30.969k i/100ms
                SHA2    17.655k i/100ms
Calculating -------------------------------------
                 MD5    402.545k (± 0.9%) i/s -      2.024M in   5.035449s
              RMD160    239.013k (± 0.7%) i/s -      1.206M in   5.049813s
                SHA1    402.324k (± 0.8%) i/s -      2.013M in   5.008906s
                SHA2    199.182k (± 0.8%) i/s -      1.006M in   5.058252s
                   with 95.0% confidence

Comparison:
                 MD5:   402545.1 i/s
                SHA1:   402323.6 i/s - same-ish: difference falls within error
              RMD160:   239013.3 i/s - 1.68x  (± 0.02) slower
                SHA2:   199182.2 i/s - 2.02x  (± 0.03) slower
                   with 95.0% confidence
```

We used a 312-byte message for the computation. At this level SHA1 and
MD5 perform nearly identical. A few ad-hoc runs with larger sizes
actually shows SHA1 out performs MD5 in terms of speed in Ruby. Given
that SHA1 is only slightly larger in terms of final hash size, is just
as fast/faster, and provides less collisions we choose it over MD5.

Circling back to the discussion of headers and the fact that this cache
implementation is not the same as the HTTP cache let's discuss our
inclusion of `cache_control`. By exposing the underlying `Cache-Control`
header like through this setting it allows an app to forward potentially
reasonable defaults (such as `max-age`) to the end client.

The following is an example of using all of this together in a Rails app
to support conditional `GET`:

```ruby
DocRepo.request(params[:slug]) do |on|
  on.complete do |doc|
    response.cache_control = doc.cache_control
    if stale?(etag: doc.cache_version, last_modified: doc.last_modified)
      @doc = doc
    end
  end
end
```

Concerns
--------

The more I look at the app side of things the less I like the fact the
underlying HTTP layer is leaking through. However, it's proving hard to
cleanly separate the two.

For example, consider the `Cache-Control` header. It's reasonable for an
app to consider mirroring the origin server's cache settings. Or at
least mirroring some of those settings such as `max-age`. This way the
final end client has some reasonable approximation of how long content
may be valid. Without a way to expose these settings that wouldn't be
possible.

I think a bit of this concern stems from having exposed both an `etag`
and now a `cache_version` attribute. These attributes have different
values and represent extremely similar concepts which may even be used
by an app in an identical manner.