Drop `DocRepo::Doc#etag` attribute
This addresses the concern from commit 631d80752983201 regarding having
potentially conflicting / confusing cache versions. Per that commit both
`cache_version` and `etag` represent very similar concepts but have
different values (ref that commit for full details).

We've accomplished a basic level of Rails integration, which is our main
target web framework at the moment, at this point. During that time
we've demonstrated that the proper way to handle E-Tag related
conditional `GET` is through the `cache_` prefixed APIs. Retaining the
`etag` attribute does not provide the app additional value. In fact it
is more likely to cause problems and lost developer time.

Since this isn't released yet it's better to just remove this today than
to have to continue to support it down stream. If this decision turns
out to be a mistake it's easy to add this back later once we have a
demonstrable reason for it.