Minor clean-up around method blocks
The `DocRepo.configure` interface is meant to take a block to customize
the configuration. Not providing the block is likely a developer error.
Instead of silently turning such a call into a no-op this raises an
error when a block isn't provided. Thus alerting the developer to their
oversight.

Conversely, this adds the ability for `DocRepo::Configuration` and
`DocRepo::ResultHandler` to yield themselves on creation when given a
block. This is a relatively standard pattern which is useful for
configuration like objects. We can consider the `ResultHandler` the
configuration for which strategies to perform given various results.