Simply loading the legacy cache includes it
This is a change made so that people who wish to force the legacy cache
after moving to Rails 5.2 can easily do so. Without this we need to
explain how to patch things in and that's not a very friendly way to
handle it. Now they can simply require the legacy file to get the
behavior back.